Move stuff into core
