+++ src/Microsoft.Extensions.Identity.Core/AuthenticatorTokenProvider.cs
