+++ src/Microsoft.Extensions.Identity.Core/EmailTokenProvider.cs
