+++ src/Microsoft.Extensions.Identity.Core/IdentityServiceCollectionExtensions.cs
@@ -1,0 +1,44 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.Extensions.DependencyInjection.Extensions;
+
+namespace Microsoft.Extensions.DependencyInjection
+{
+    /// <summary>
+    /// Contains extension methods to <see cref="IServiceCollection"/> for configuring identity services.
+    /// </summary>
+    public static class IdentityServiceCollectionExtensions
+    {
+        /// <summary>
+        /// Adds and configures the identity system for the specified User and Role types.
+        /// </summary>
+        /// <typeparam name="TUser">The type representing a User in the system.</typeparam>
+        /// <param name="services">The services available in the application.</param>
+        /// <param name="setupAction">An action to configure the <see cref="IdentityOptions"/>.</param>
+        /// <returns>An <see cref="IdentityBuilder"/> for creating and configuring the identity system.</returns>
+        public static IdentityBuilder AddIdentityCore<TUser>(this IServiceCollection services, Action<IdentityOptions> setupAction)
+            where TUser : class
+        {
+            // Services used by identity
+            services.TryAddScoped<IUserValidator<TUser>, UserValidator<TUser>>();
+            services.TryAddScoped<IPasswordValidator<TUser>, PasswordValidator<TUser>>();
+            services.TryAddScoped<IPasswordHasher<TUser>, PasswordHasher<TUser>>();
+            services.TryAddScoped<ILookupNormalizer, UpperInvariantLookupNormalizer>();
+            // No interface for the error describer so we can add errors without rev'ing the interface
+            services.TryAddScoped<IdentityErrorDescriber>();
+            services.TryAddScoped<IUserClaimsPrincipalFactory<TUser>, UserClaimsPrincipalFactory<TUser>>();
+            services.TryAddScoped<UserManager<TUser>, UserManager<TUser>>();
+
+            if (setupAction != null)
+            {
+                services.Configure(setupAction);
+            }
+
+            return new IdentityBuilder(typeof(TUser), services);
+        }
+    }
+}
