+++ src/Microsoft.Extensions.Identity.Core/PasswordHasherOptions.cs
