+++ src/Microsoft.Extensions.Identity.Core/PasswordValidator.cs
