+++ src/Microsoft.Extensions.Identity.Core/TotpSecurityStampBasedTokenProvider.cs
