+++ src/Microsoft.Extensions.Identity.Core/UpperInvariantLookupNormalizer.cs
