+++ src/Microsoft.Extensions.Identity.Core/UserValidator.cs
